Refine scalable recommendation frameworks

+++ trunk/GenericSystem/genericsystem-common-algorithm/src/main/java/edu/tongji/data/SparseMatrix.java
@@ -826,9 +826,9 @@ public class SparseMatrix implements Serializable {
      * @param isRow
      * @return
      */
-    public float[][] probability(int[] rows, int[] cols, double maxValue, double minValue,
-                                 boolean isRow) {
-        float[][] probs = null;
+    public double[][] probability(int[] rows, int[] cols, double maxValue, double minValue,
+                                  boolean isRow) {
+        double[][] probs = null;
         int weightSize = Double.valueOf(maxValue / minValue).intValue();
 
         if (rows == null && cols == null) {
@@ -852,7 +852,7 @@ public class SparseMatrix implements Serializable {
             }
 
             //construct probabilities with Lagrange Smoothing
-            probs = new float[M][0];
+            probs = new double[M][0];
 
             //compute the probabilities 
             for (int row : rows) {
@@ -860,8 +860,8 @@ public class SparseMatrix implements Serializable {
                 int[] indexList = Fu.indexList();
 
                 //count
-                float[] localPro = new float[weightSize];
-                Arrays.fill(localPro, 1.0f);
+                double[] localPro = new double[weightSize];
+                Arrays.fill(localPro, 1.0d);
 
                 int total = weightSize;
                 if (indexList != null) {
@@ -891,7 +891,7 @@ public class SparseMatrix implements Serializable {
             }
 
             //construct probabilities with Lagrange Smoothing
-            probs = new float[N][0];
+            probs = new double[N][0];
 
             //compute the probabilities
             for (int col : cols) {
@@ -899,8 +899,8 @@ public class SparseMatrix implements Serializable {
                 int[] indexList = Gi.indexList();
 
                 //count
-                float[] localPro = new float[weightSize];
-                Arrays.fill(localPro, 1.0f);
+                double[] localPro = new double[weightSize];
+                Arrays.fill(localPro, 1.0d);
 
                 int total = weightSize;
                 if (indexList != null) {
